requirements.txt
configue==3.0.5
dill==0.3.6
fire==0.5.0
lightgbm==3.3.5
lightning-utilities==0.3.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
numpy==1.23.5
optuna==3.1.0
pandas==1.5.3
pytorch-lightning==1.8.1
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
statsforecast==0.5.5
statsmodels==0.13.5
sktime==0.15.1
torch==1.13.1
torchmetrics==0.11.0
torchvision==0.14.1
u8darts==0.20.0
xarray==2023.1.0
netCDF4==1.6.2
parameterized==0.8.1
